+++ CHANGES.rst
@@ -41,6 +41,7 @@ in progress
 - Refactoring: Add ``eskema.io.open`` as a wrapper around ``fsspec.open``
 - Refactoring: Add ``eskema.io.to_dataframe`` from ``eskema.autopk``
 - Refactoring: Add ``eskema.util`` folder instead of single ``util.py``
+- UX: Add ``eskema info`` subcommand
 
 
 2023-01-xx 0.0.0